Enable video streaming using AP1302 and AR1335 (#6)
* Added Linux firmware loader daemon and the AP1302 FW bin file
The firmware daemon recipe was taken from a new version of meta-openembedded.
It's not available in our current Yocto release. The AP1302 driver needs this
daemon as it enables the driver to load firmware during the driver probe.

* Added U-Dev rules for modprobe of AP1302
AP1302 has some configuration options which need to
be set at load time. 

* Added media controller pipeline setup script for our sensor configuration
At the moment this just works for 1 sensor. But we can modify this later
to work for multiple sensors. Or we can just can setup the pipeline
in code.