Feature/add-gstreamer-plugins (#21)
* Added GStreamer plugins to the core-image include OpenCVA
OpenCVA adds a bin file which must be loaded in U-Boot to a
specific address. I am also using the AI-SDK 4.0.0 release
which includes the H264 Encoder plugin inside
meta-rz-features/meta-rz-codecs. I'll need to add these
seperately as they are a behind licensing agreement
on Renesas's website.

* Updated v4l2-init script to setup pipelines on user selected device
Also updated our layer conf to include meta-rz-codecs, which is
externally downloaded and meta-rz-opencva which provides
access to the DRP via OpenCV (ie the OpenCV accelerator).

* Disable v4l2-init service from auto starting