Bugfix for Yocto Build Error due to include swupdate
Build was throwing warnings like:
WARNING: preferred version v2021.10% of u-boot not available (for item u-boot-default-env)
WARNING: versions of u-boot available: 1:2020.01
WARNING: preferred version v2021.10% of u-boot not available (for item u-boot-env)
WARNING: versions of u-boot available: 1:2020.01

And also throwing this bitbake error:
ERROR: Multiple versions of u-boot are due to be built

I think this is because u-boot-default-env creates has some dependencies
for certain versions of U-Boot which we don't have in our build. Dimitrios had
included his fix to remove override RRECOMMENDS_${PN}-bin_append_class-target.
But this wasn't actually overriding the string, it was just appending to it. We actually
want to remove u-boot-default-env from the list.